DYN-9707 rationalize graph lock prompt
